pyproject.toml
[tool.poetry]
name = "powercord"
version = "0.1.0"
description = "A server framework for advanced Discord bot functionality"
authors = ["Arthur Davis <arthur@hiphidelity.com>"]
license = "The MIT License"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = { version = "^0.116.1", extras = ["standard"] }
google-cloud-logging = "^3.12.1"
google-cloud-secret-manager = "^2.24.0"
nextcord = { version = "^3.1.0", extras = ["voice"] }
pathvalidate = "^3.3.1"
pg8000 = "^1.31.4"
python-fasthtml = "^0.12.21"
sanitize-filename = "^1.2.0"
SQLAlchemy = "^2.0.41"
uvicorn = { version = "^0.35.0", extras = ["standard"] }
sqlmodel = "^0.0.33"
alembic = "^1.18.4"
fastsql = "^2.1.2"
psutil = "^7.2.2"

[tool.poetry.group.test.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: marks tests as unit tests (deselect with '-m \"not unit\"')",
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
omit = [
    "app/extensions/example/*",
    "app/bot/views.py",
    "app/main_bot.py",
    "app/main_api.py",
    "app/main_ui.py",
    "app/ui/dashboard.py",
    "app/ui/helpers.py",
]


[tool.poetry.group.dev.dependencies]
ruff = "^0.15.0"
bandit = "^1.9.3"
mypy = "^1.19.1"
pre-commit = "^4.5.1"
types-requests = "^2.32.4.20260107"
playwright = "^1.58.0"

[tool.pyrefly]
project-includes = ["app"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
    "S", # bandit
]
ignore = [
    "E501", # line too long, handled by black
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"app/ui/*.py" = ["F403", "F405"]
"app/main_ui.py" = ["F403", "F405"]
"app/**/*.py" = ["E402"]
"tests/**/*.py" = ["S"]
"tests/**" = ["S"]
"**/tests/**" = ["S"]
"app/ui/auth.py" = ["S105"]
"app/common/gsm_loader.py" = ["S607"]
"app/extensions/**/*.py" = ["F403", "F405", "S311", "B008", "E741", "S101"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "app/extensions/example",
    "app/ui",
    "app/main_ui.py",
    "app/main_bot.py",
    "app/bot/powerloader.py",
    "alembic",
]


[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]

[tool.bandit.per_file_ignores]
"app/common/gsm_loader.py" = ["B404", "B603", "B607"]
"app/extensions/example/cog.py" = ["B311"]
"app/ui/auth.py" = ["B105"]
